+++ nv/vi/core.py
@@ -15,11 +15,11 @@
 # You should have received a copy of the GNU General Public License
 # along with NeoVintageous.  If not, see <https://www.gnu.org/licenses/>.
 
-from sublime import DRAW_NO_FILL
 from sublime import set_timeout
 import sublime_plugin
 
 from NeoVintageous.nv.state import State
+from NeoVintageous.nv.ui import ui_region_flags
 from NeoVintageous.nv.vi.utils import IrreversibleTextCommand
 from NeoVintageous.nv.vim import console_message
 
@@ -102,16 +102,23 @@ class ViCommandMixin(object):
         state.xpos = xpos
 
     def outline_target(self):
-        # TODO rework 'vintageous_visualyank' feature to be more similar
-        # to https://github.com/machakann/vim-highlightedyank
-        # See http://vimcasts.org/episodes/neovim-eyecandy/#shownotes
+        view = self._view
+        _get = view.settings().get
 
-        if not self._view.settings().get('vintageous_visualyank'):
+        if not _get('highlightedyank'):
             return
 
-        sels = list(self._view.sel())
-        set_timeout(lambda: self._view.erase_regions('vi_yy_target'), 350)
-        self._view.add_regions('vi_yy_target', sels, 'comment highlighted.yank', '', DRAW_NO_FILL)
+        view.add_regions(
+            'highlightedyank',
+            list(view.sel()),
+            scope='string highlightedyank',
+            flags=ui_region_flags(_get('highlightedyank_style'))
+        )
+
+        set_timeout(
+            lambda: view.erase_regions('highlightedyank'),
+            _get('highlightedyank_duration')
+        )
 
 
 # TODO [refactor] Move to commands module